Refactor PDS Goal Calculator components to implement step navigation and improve layout
