+++ src/components/Reports/PdsGoalCalculator/ReimbursableExpenses/ReimbursableExpensesStep.test.tsx
@@ -4,14 +4,13 @@ import { PdsGoalCalculatorTestWrapper } from '../PdsGoalCalculatorTestWrapper';
 import { ReimbursableExpensesStep } from './ReimbursableExpensesStep';
 
 describe('ReimbursableExpensesStep', () => {
-  it('renders the monthly and annual reimbursable expenses nav items', () => {
+  it('renders the reimbursable expenses heading', () => {
     const { getByText } = render(
       <PdsGoalCalculatorTestWrapper>
         <ReimbursableExpensesStep />
       </PdsGoalCalculatorTestWrapper>,
     );
 
-    expect(getByText('Monthly Reimbursable Expenses')).toBeInTheDocument();
-    expect(getByText('Annual Reimbursable Expenses')).toBeInTheDocument();
+    expect(getByText('Reimbursable Expenses')).toBeInTheDocument();
   });
 });
